Add check for ElementInterface
Check if setting record exist before inserting

+++ src/services/handlers/PublishDraftHandler.php
@@ -104,6 +104,10 @@ class PublishDraftHandler
                         // Clear current neo field value
                         $neoField = $draftElementLanguageToUpdate->getFieldValue($field->handle);
                         foreach ($neoField as $block) {
+                            if(!$block instanceof ElementInterface) {
+                                continue;
+                            }
+
                             Craft::$app->getElements()->deleteElement($block);
                         }
                         Craft::$app->getElements()->saveElement($draftElementLanguageToUpdate);
